+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.6.0](https://github.com/Ranga-Schooling/trustai-marketplace/compare/v1.5.0...v1.6.0) (2026-08-06)
+
+
+### Features
+
+* **app:** merge promo and intro into one landing column; wire up his… ([#29](https://github.com/Ranga-Schooling/trustai-marketplace/issues/29)) ([9ebe90a](https://github.com/Ranga-Schooling/trustai-marketplace/commit/9ebe90a410423f26186b2798c1889e1a69f85e00))
+
+## [1.5.0](https://github.com/Ranga-Schooling/trustai-marketplace/compare/v1.4.0...v1.5.0) (2026-08-06)
+
+
+### Features
+
+* align sign in/register screen with approved wireframes (US-1.1/… ([#25](https://github.com/Ranga-Schooling/trustai-marketplace/issues/25)) ([707f2e4](https://github.com/Ranga-Schooling/trustai-marketplace/commit/707f2e4778de8b3b9c798f2d210b72d41b65ef1e)), closes [#20](https://github.com/Ranga-Schooling/trustai-marketplace/issues/20) [#20](https://github.com/Ranga-Schooling/trustai-marketplace/issues/20) [#20](https://github.com/Ranga-Schooling/trustai-marketplace/issues/20)
+
+## [1.4.0](https://github.com/Ranga-Schooling/trustai-marketplace/compare/v1.3.0...v1.4.0) (2026-08-06)
+
+
+### Features
+
+* view and edit account profile (US-1.4) ([#24](https://github.com/Ranga-Schooling/trustai-marketplace/issues/24)) ([5c3c918](https://github.com/Ranga-Schooling/trustai-marketplace/commit/5c3c918e837646b55156f4d51a9b40df4ee9fa19))
+
 ## [1.3.0](https://github.com/Ranga-Schooling/trustai-marketplace/compare/v1.2.1...v1.3.0) (2026-08-02)